Corrected definition of Kostka number
Corrected that the definition of Kostka number.  It wasn't necessarily wrong before, but you had to interpret the second parameter in a nonstandard way to get the Kostka numbers.  That has now been fixed.

Things have also been reorganized to have a better separation of distinct concepts into distinct files.